backend/requirements.txt
flask
flask-cors
flask-login
flask_sqlalchemy
requests
beautifulsoup4
python-dotenv
werkzeug